# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Content System API",
        "description": "API для системы комментариев к контенту",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
        "/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Получение списка комментариев",
                "operationId": "e21e2ae66afe62e7d8d0b023798c70af",
                "parameters": [
                    {
                        "name": "commentable_type",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "news",
                                "video-post"
                            ]
                        }
                    },
                    {
                        "name": "commentable_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "nullable": true
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "description": "ID последнего комментария предыдущей страницы",
                            "type": "integer",
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentList"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Создание комментария",
                "operationId": "478e0f4964ebbe8d3a22df34f4fcddab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "commentable_type": {
                                        "type": "string",
                                        "enum": [
                                            "news",
                                            "video-post"
                                        ],
                                        "example": "news"
                                    },
                                    "commentable_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Текст комментария"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Комментарий создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/comments/{id}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Получение комментария",
                "operationId": "c2aac1315d33473f4b25ddf45941b871",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено"
                    }
                }
            },
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Обновление комментария",
                "operationId": "0904942451844b2582af15f7c38988b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Обновленный текст комментария"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Комментарий обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Удаление комментария",
                "operationId": "dc71fcedd4e6d8fe9a74a965257cc90b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Комментарий удален"
                    },
                    "404": {
                        "description": "Не найдено"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Comments"
                ],
                "summary": "Обновление комментария",
                "operationId": "9e2cb38309186fe90d8552f3bd6e8b63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Обновленный текст комментария"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Комментарий обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CommentList": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comment"
                        }
                    },
                    "limit": {
                        "type": "integer",
                        "example": 10
                    },
                    "next_cursor": {
                        "description": "ID последнего комментария для следующей страницы",
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    },
                    "has_more": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "Comment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "commentable_type": {
                        "type": "string",
                        "example": "App\\Models\\News"
                    },
                    "commentable_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "content": {
                        "type": "string",
                        "example": "Текст комментария"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "User Name"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}